apps/api/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Caching (optional - comment out if not using Redis)
# redis==5.0.1

# Timezone handling
pytz==2024.1

# Geocoding
geopy==2.4.1

# HTTP Client
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Templates
jinja2==3.1.3

# Date/Time
python-dateutil==2.8.2

# Utilities
orjson==3.9.12
python-dotenv==1.0.1
